hot fix combobox image type in animation
cannot remove items from combobox, when the n-th index is used.
when exr was disabled and tiff was enabled the combobox yielded exr
type, when tiff was selected.

TODO: for next release make new type comboBoxId which has keys already
applied in ui file, this way we can manipulate the combobox without
loosing relation to the key.
Plus: This will also make the indices named in the settings file.
backwards compatibility could be:
if int parccelable -> value is index
else value lookup for index