netrender - connect clients while rendering
this should fix issue #90
first tests seem to be fine.

The is no SendJob function anymore, but SetJob.
This behaves as before and sends the job to all connected clients, but
also stores the job in netrender persistent var 'msgCurrentJob'
when a new client connects successfully and msgCurrentJob is set, this
gets send to the new client.

@buddhi please double check the code in netrender.cpp