Volume slicing of fractal for simple 3d reconstruction
new class volume which receives a bounding box and a x-y-z resolution.
it calculates for every voxel inside the volume if it is inside the
fractal or not.
volume data will be stored as PNG image slices in black and white.

here is a little testing code:

```
cVolume vol(50, 50, 50, CVector3(-1, -1, -1), CVector3(1, 1, 1));
	vol.CalculateVolume();
	vol.Store("/home/sebastian/test/");
```

TODO:
- make UI and connect to class
- update progress
- performance enhance