added thumbnail option to render all example thumbnails
added new button under preferencesdialog to 'generate' thumbnails
this will take all example settings and render the thumbnails to the
cache.
This is intended to be used to generate the cache as a developer and
provide it on a server for loading via 'load'
anyway users should feel free to use it themselves. corresponding info
text onclick on button