added image_save_dialog to select image quality in save dialog
See also issue #103.

These image quality adjustments will be used for the image saving and be
stored in the global preferences.
I would still keep image settings in preferences, since the user might
only want to change these values for animation, where there is no
interception with a dialog before saving.

Drawback is: Now we are keeping the ui for image quality settings
duplicate.

Any ideas how to rearrange this are welcome

BTW: Also added a fix for jpg saving: When the normals were not rendered
and the image saved as jpg with normal channel the program crashed (null
pointer dereference), now the program only prints a warning.