added gamepad keymap
The gamepad dock now contains a tooltip with a visual gamepad keymap.
The keymap is based on nicefrog's design (http://opengameart.org/
content/generic-gamepad-template) and present in the file "gamepad Keymap
source.svg" changes and explanations for mandelbulber made with
inkscape.
Since displaying of svg in inkscape and qt build vary (arrows are
missing) the svg is exported as an png. This is only the output, changes
should only be made to the svg and exported again.

This deals with issue #144 @mancoast please check, if this is correct
and complete.