new feature for auto limit - SetBoundingBoxAsLimits
SetBoundingBoxAsLimits - Calculates the bounding box of the fractal and
sets these as the limit values. This can be helpful to automatically set
the limits for the voxel export.

The bounding box is calculated in the following way:
foreach limit dimension a virtual plane is placed far away of the
fractal orthogonal to the limit direction.
The distance of the plane to the fractal is calculated and this distance
is then used to determine the limit value (plus some little border).

To calculate a distance of a plane to the fractal the new method
CalculateDistanceMinPlane is used.
CalculateDistanceMinPlane takes the plane and iteratively moves the
plane point closer to the fractal surface by moving along the plane
orthogonal and slightly moving orthogonal to the plane orthogonal in any
direction which has a minimum of fractal distance in each marching step.

This is not an universal solution but works quite okay for some test
cases (try menger sponge, quaternion).

TODO: user defined bounding box max value