custom window settings - remove now with new action and window list
gnome is not capable of displaying custom layouts in a qmenu, this is
why there can be only one signal for an action in a qmenu.
custom window settings can now be removed with an extra action
"actionRemove_Window_settings" this opens a popup with a combobox of the
existing custom layouts to select one for deletion, as buddhi1980
suggested.