Bugfix FilePathHelper
If the path was not found, the same path was tried again and again.
Fix this by trying the alternate paths instead.

Add newlines to the printouts to not print all paths on the line.

The changes makes it possible to run the example file

	mandelbulber2/deploy/share/mandelbulber2/examples/Robert Pancoast
	collection - license Creative Commons (CC-BY 4.0)/icoastahedron.fract

even if the repo isn't checked out to D:/GIT/mandelbulber2.