cimage index access - check by assert
every access to a cimage index (access to a pixel from any image buffer)
should be proper in the surrounding context (in range). In Debug mode
this will now get checked by an assert. In Release mode this check will
be off, so this error shall never occur in Release.
I have checked some settings in GUI mode and ran the mandelbulber --test
If you get this assertion somewhere in the code, debug and fix the
faulting context.