removed "dont override"-function from formula cl files
At the time the cpp-opencl-transpiler was written, it looked like it may
not fully be able to autogenerate all opencl-formula files. This is why
a mechanism has been placed in the files, to prevent override which
would enable writing custom opencl code into the file to fix issues or
make it run faster.

Now it is clear that all formulas can be autogenerated by this
transpiler in an almost optimal way, so the custom file - function is
not needed. This is now presented in the cl file header.